+++ src/be/nikiroo/fanfix/supported/BasicSupport.java
@@ -156,13 +156,13 @@ public abstract class BasicSupport {
 	private URL currentReferer; // with only one 'r', as in 'HTTP'...
 
 	// quote chars
-	private char openQuote = Instance.getTrans().getChar(
+	private char openQuote = Instance.getTrans().getCharacter(
 			StringId.OPEN_SINGLE_QUOTE);
-	private char closeQuote = Instance.getTrans().getChar(
+	private char closeQuote = Instance.getTrans().getCharacter(
 			StringId.CLOSE_SINGLE_QUOTE);
-	private char openDoubleQuote = Instance.getTrans().getChar(
+	private char openDoubleQuote = Instance.getTrans().getCharacter(
 			StringId.OPEN_DOUBLE_QUOTE);
-	private char closeDoubleQuote = Instance.getTrans().getChar(
+	private char closeDoubleQuote = Instance.getTrans().getCharacter(
 			StringId.CLOSE_DOUBLE_QUOTE);
 
 	/**